New subdirectory grammar_test created. All package files moved there. Unit tests have their import statements changed according to new path to the files.
