+++ setup.py
@@ -6,7 +6,7 @@ setup(name='opencog-ull',
       author='Opencog ULL Team',
       url='http://github.com/singnet/language-learning',
       packages=['grammartest'],
-      package_dir={'grammartest': 'src/link_grammar'},
+      package_dir={'grammartest': 'src/grammar_test'},
       platform='any',
       license='MIT',
       classifiers=[
