+++ src/grammar_test/README.md
@@ -1,0 +1,7 @@
+# `grammartest` package 
+
+This subdirectory contains several library files which make up `grammartest` package. 
+As the name suggests, the purpose of the package is to test and evaluate induced grammar.
+
+
+
