+++ tests/test_grammartester.py
@@ -1,8 +1,11 @@
 import unittest
-from link_grammar.grammartester import GrammarTester
-from link_grammar.lginprocparser import LGInprocParser
-from link_grammar.parsemetrics import ParseMetrics
-from link_grammar.optconst import *
+from grammar_test.grammartester import GrammarTester, test_grammar
+from grammar_test.lginprocparser import LGInprocParser
+from grammar_test.parsemetrics import ParseMetrics
+from grammar_test.optconst import *
+
+import cProfile
+import pstats
 
 # dict = "/usr/local/share/link-grammar/en"
 # dict = "en"
@@ -25,13 +28,21 @@ from link_grammar.optconst import *
 tmpl = "/home/alex/data/dict/poc-turtle"
 grmr = "/home/alex/data/dict"
 limit = 100
-opts = BIT_NO_LWALL | BIT_NO_PERIOD | BIT_STRIP | BIT_RM_DIR | BIT_LOC_LANG | BIT_PARSE_QUALITY | BIT_ULL_IN #| BIT_OUTPUT_DIAGRAM #| BIT_ULL_IN | BIT_SEP_STAT
+opts = BIT_NO_LWALL | BIT_NO_PERIOD | BIT_STRIP | BIT_RM_DIR | BIT_DPATH_CREATE | BIT_LOC_LANG | BIT_PARSE_QUALITY #| BIT_ULL_IN #| BIT_OUTPUT_DIAGRAM #| BIT_ULL_IN | BIT_SEP_STAT
 
 # # Gutenberg Children Parses
 # dict = "en"
 # corp = "/home/alex/data/corpora/cleaned_Gutenberg_Children"
-# dest = "/home/alex/data2/parses/cleaned_Gutenberg_Children"
+# dest = "/home/alex/data2/parses/cleaned_Gutenberg_Children_ref"
 # ref = None
+
+
+# # Gutenberg Children Parses
+# dict = "/home/alex/data2/parses/Gutenberg-Children-2018-05-30"
+# corp = "/home/alex/data/corpora/cleaned_Gutenberg_Children"
+# dest = "/home/alex/data2/parses/Gutenberg-Children-2018-05-30"
+# ref = "/home/alex/data2/parses/cleaned_Gutenberg_Children_ref"
+# # ref = None
 
 
 # ref = "/home/alex/data/corpora"
@@ -49,7 +60,7 @@ opts = BIT_NO_LWALL | BIT_NO_PERIOD | BIT_STRIP | BIT_RM_DIR | BIT_LOC_LANG | BI
 # # ref = None
 
 # AGI-2018 Test
-dict = "/home/alex/data/parses/AGI-2018-paper-data-2018-04-22/POC-English-NoAmb-LEFT-WALL+period"
+dict = "/home/alex/data2/parses/AGI-2018-paper-data-2018-04-22/POC-English-NoAmb-LEFT-WALL+period"
 corp = "/home/alex/data/poc-english/poc_english_noamb.txt"
 dest = "/home/alex/data2/parses/AGI-2018-paper-data-2018-04-22/POC-English-NoAmb-LEFT-WALL+period"
 ref = "/home/alex/data/poc-english/poc_english_noamb_parse_ideal.txt"
@@ -67,6 +78,15 @@ ref = "/home/alex/data/poc-english/poc_english_noamb_parse_ideal.txt"
 # ref = "/home/alex/data2/parses/Gutenberg-Alice-2018-06-01/parses/alice_11-0_txt_split_default.txt.ull"
 # # ref = None
 
+# # Child Directed Speech
+# dict = "en"
+# # dict = "/home/alex/data2/parses/Gutenberg-Alice-2018-06-01"
+# corp = "/home/alex/data/corpora/ChildDirectedSpeech"
+# dest = "/home/alex/data2/parses/ChildDirectedSpeech"
+# # ref = "/home/alex/data2/parses/Gutenberg-Alice-2018-06-01/parses/alice_11-0_txt_split_default.txt.ull"
+# ref = None
+
+
 
 # # PubMed-2018-06-01 parse for ULL reference
 # dict = "en"
@@ -76,6 +96,8 @@ ref = "/home/alex/data/poc-english/poc_english_noamb_parse_ideal.txt"
 
 
 class GrammarTesterTestCase(unittest.TestCase):
+
+    # @unittest.skip
     def test_test(self):
         pr = LGInprocParser()
 
@@ -86,6 +108,21 @@ class GrammarTesterTestCase(unittest.TestCase):
 
         # self.assertEqual(25, gt._total_dicts)
         self.assertEqual(88, pm.sentences)
+
+    @unittest.skip
+    def test_test_grammar(self):
+        file_name = "/var/tmp/test-grammar-stats"
+
+        def run_test():
+            test_grammar(corp, dest, dict, grmr, tmpl, limit, opts, ref)
+            return ""
+
+        cProfile.run(run_test())
+
+        # cProfile.run(print("!!!!!!"))
+        # p = pstats.Stats(file_name)
+        # p.strip_dirs().sort_stats(-1).print_stats()
+        self.assertTrue(True)
 
 if __name__ == '__main__':
     unittest.main()
